Reorder base Data types to keep aiida.orm.data.base import
The old path to import the base types Bool, Float, Int and Str
was `aiida.orm.data.base`. To keep backwards compatibility we
need to import the classes in this file from their new individual
module files. However, that means that the BaseType class needs
to be moved out of that file, in order to prevent a circular
import dependency. Moving it to the __init__.py file of the
aiida.orm.data module works
