+++ aiida/backends/tests/base_dataclasses.py
@@ -13,8 +13,8 @@ import operator
 from aiida.backends.testbase import AiidaTestCase
 from aiida.common.exceptions import ModificationNotAllowed
 from aiida.orm import load_node
-from aiida.orm.data.base import NumericType
-from aiida.orm.data.base import List
+from aiida.orm.data.numeric import NumericType
+from aiida.orm.data import List
 from aiida.orm.data.bool import Bool, get_true_node, get_false_node
 from aiida.orm.data.float import Float
 from aiida.orm.data.int import Int
