+++ aiida/orm/data/__init__.py
@@ -7,11 +7,26 @@
 # For further information on the license, see the LICENSE.txt file        #
 # For further information please visit http://www.aiida.net               #
 ###########################################################################
+from abc import ABCMeta
+from collections import MutableSequence
+
+try:
+    from functools import singledispatch
+except ImportError:
+    from singledispatch import singledispatch
+
 from aiida.orm.node import Node
 from aiida.common.links import LinkType
 from aiida.common.lang import override
 from aiida.common.exceptions import ModificationNotAllowed
 
+
+@singledispatch
+def to_aiida_type(value):
+    """
+    Turns basic Python types (str, int, float, bool) into the corresponding AiiDA types.
+    """
+    raise TypeError("Cannot convert value of type {} to AiiDA type.".format(type(value)))
 
 
 class Data(Node):
@@ -369,3 +384,183 @@ class Data(Node):
         ##     raise ValidationError("License of the object ({}) requires "
         ##                           "attribution, while none is given in the "
         ##                           "description".format(self.source['license']))
+
+
+
+class BaseType(Data):
+    """
+    Store a base python type as a AiiDA node in the DB.
+
+    Provide the .value property to get the actual value.
+    """
+    __metaclass__ = ABCMeta
+
+    def __init__(self, *args, **kwargs):
+        try:
+            getattr(self, '_type')
+        except AttributeError:
+            raise RuntimeError("Derived class must define the _type class member")
+
+        super(BaseType, self).__init__(**self._create_init_args(*args, **kwargs))
+
+    def set_typevalue(self, typevalue):
+        _type, value = typevalue
+        self._type = _type
+        if value:
+            self.value = value
+        else:
+            self.value = _type()
+
+    @property
+    def value(self):
+        return self.get_attr('value')
+
+    @value.setter
+    def value(self, value):
+        self._set_attr('value', self._type(value))
+
+    def __str__(self):
+        return self.value.__str__()
+
+    def __repr__(self):
+        return self.value.__repr__()
+
+    def __eq__(self, other):
+        if isinstance(other, BaseType):
+            return self.value == other.value
+        else:
+            return self.value == other
+
+    def __ne__(self, other):
+        if isinstance(other, BaseType):
+            return self.value != other.value
+        else:
+            return self.value != other
+
+    def new(self, value=None):
+        return self.__class__(typevalue=(self._type, value))
+
+    def _create_init_args(self, *args, **kwargs):
+        if args:
+            assert not kwargs, "Cannot have positional arguments and kwargs"
+            assert len(args) == 1, \
+                "Simple data can only take at most one positional argument"
+
+            kwargs['typevalue'] = (self._type, self._type(args[0]))
+
+        elif 'dbnode' not in kwargs:
+            if 'typevalue' in kwargs:
+                assert kwargs['typevalue'][0] is self._type
+                if kwargs['typevalue'][1] is not None:
+                    kwargs['typevalue'] = \
+                        (self._type, self._type(kwargs['typevalue'][1]))
+            else:
+                kwargs['typevalue'] = (self._type, None)
+
+        else:
+            assert len(kwargs) == 1, \
+                "When specifying dbnode it can be the only kwarg"
+
+        return kwargs
+
+
+
+class List(Data, MutableSequence):
+    """
+    Class to store python lists as AiiDA nodes
+    """
+    _LIST_KEY = 'list'
+
+    def __getitem__(self, item):
+        return self._get_list()[item]
+
+    def __setitem__(self, key, value):
+        l = self._get_list()
+        l[key] = value
+        if not self._using_list_reference():
+            self._set_list(l)
+
+    def __delitem__(self, key):
+        l = self._get_list()
+        del l[key]
+        if not self._using_list_reference():
+            self._set_list(l)
+
+    def __len__(self):
+        return len(self._get_list())
+
+    def __str__(self):
+        return self._get_list().__str__()
+
+    def append(self, value):
+        l = self._get_list()
+        l.append(value)
+        if not self._using_list_reference():
+            self._set_list(l)
+
+    def extend(self, L):
+        l = self._get_list()
+        l.extend(L)
+        if not self._using_list_reference():
+            self._set_list(l)
+
+    def insert(self, i, value):
+        l = self._get_list()
+        l.insert(i, value)
+        if not self._using_list_reference():
+            self._set_list(l)
+
+    def remove(self, value):
+        del self[value]
+
+    def pop(self, **kwargs):
+        l = self._get_list()
+        l.pop(**kwargs)
+        if not self._using_list_reference():
+            self._set_list(l)
+
+    def index(self, value):
+        return self._get_list().index(value)
+
+    def count(self, value):
+        return self._get_list().count(value)
+
+    def sort(self, cmp=None, key=None, reverse=False):
+        l = self._get_list()
+        l.sort(cmp, key, reverse)
+        if not self._using_list_reference():
+            self._set_list(l)
+
+    def reverse(self):
+        l = self._get_list()
+        l.reverse()
+        if not self._using_list_reference():
+            self._set_list(l)
+
+    def _get_list(self):
+        try:
+            return self.get_attr(self._LIST_KEY)
+        except AttributeError:
+            self._set_list(list())
+            return self.get_attr(self._LIST_KEY)
+
+    def _set_list(self, list_):
+        if not isinstance(list_, list):
+            raise TypeError("Must supply list type")
+        self._set_attr(self._LIST_KEY, list_)
+
+    def _using_list_reference(self):
+        """
+        This function tells the class if we are using a list reference.  This
+        means that calls to self.get_list return a reference rather than a copy
+        of the underlying list and therefore self._set_list need not be called.
+        This knwoledge is essential to make sure this class is performant.
+
+        Currently the implementation assumes that if the node needs to be
+        stored then it is using the attributes cache which is a reference.
+
+        :return: True if using self._get_list returns a reference to the
+            underlying sequence.  False otherwise.
+        :rtype: bool
+        """
+        return not self.is_stored
